waiter performance improvements
inline all of the fil_waiter stuff via header, no longer requiring
the creation of a python object.

Fix all uses of cond_*wait to use a helper that only blocks for
a short period of time to allow signals to be checked. This fixes
^C not actually doing anything because of being blocked in
pthread_cond_*wait().

We currently will check signals every 250ms, but maybe we'll want to
check every 100ms or something.